refactor: move LSP-related utils into lsp package
